placemark for not zg object

+++ app/models/map_projects/map_placemark.rb
@@ -2,11 +2,11 @@ include ImageHelper
 
 class MapPlacemark < ActiveRecord::Base
   attr_accessor :related_items
-  attr_accessible :title, :map_layer_id, :related_items
+  attr_accessible :title, :map_layer_id, :related_items, :latitude, :longitude, :url, :address,
+                  :image
 
-  validates_presence_of :map_layer_id, :related_items
-  after_save :parse_related_item
-  before_save :set_address
+  validates_presence_of :map_layer_id
+  before_save :parse_related_items
 
   belongs_to :map_layer
 
@@ -16,7 +16,11 @@ class MapPlacemark < ActiveRecord::Base
   has_many :reviews,        :through => :relations, :source => :slave, :source_type => Review
   has_many :photogalleries, :through => :relations, :source => :slave, :source_type => Photogallery
 
-  def parse_related_item
+  has_attached_file :image, :storage => :elvfs, :elvfs_url => Settings['storage.url']
+
+  def parse_related_items
+    return if related_items.nil?
+
     relations.destroy_all
 
     related_items.each do |item|
@@ -28,27 +32,25 @@ class MapPlacemark < ActiveRecord::Base
 
       relation.save
     end
-  end
 
-  def set_address
-    related_items.each do |item|
-      item_type, item_id = item.split("_")
-      class_item = item_type.classify.constantize.find(item_id)
-      if class_item.is_a? Organization
-        self.title = class_item.title
-        self.latitude = class_item.latitude
-        self.longitude = class_item.longitude
-        self.image_url = resized_image_url(class_item.logotype_url, 190, 190)
-        self.address = class_item.address.to_s
-        self.url = "/organizations/#{class_item.slug}"
+    relations.each do |relation|
+      relation = relation.slave
+      if relation.is_a? Organization
+        self.title = relation.title
+        raise self.title
+        self.latitude = relation.latitude
+        self.longitude = relation.longitude
+        self.image_url = resized_image_url(relation.logotype_url, 190, 190)
+        self.address = relation.address.to_s
+        self.url = "/organizations/#{relation.slug}"
       else
-        self.title = class_item.title
-        self.latitude = class_item.organization.latitude
-        self.longitude = class_item.organization.longitude
-        self.image_url = resized_image_url(class_item.poster_url, 190, 260)
-        self.address = class_item.address
-        self.when = "#{item_type.classify}Decorator".constantize.new(class_item).human_when
-        self.url = "/#{item_type}/#{class_item.slug}"
+        self.title = relation.title
+        self.latitude = relation.organization.latitude
+        self.longitude = relation.organization.longitude
+        self.image_url = resized_image_url(relation.poster_url, 190, 260)
+        self.address = relation.address
+        self.when = "#{relation.class.name}Decorator".constantize.new(relation).human_when
+        self.url = "/afisha/#{relation.slug}"
       end
     end
   end
@@ -58,16 +60,19 @@ end
 #
 # Table name: map_placemarks
 #
-#  id           :integer          not null, primary key
-#  title        :string(255)
-#  latitude     :float
-#  longitude    :float
-#  image_url    :string(255)
-#  url          :string(255)
-#  address      :string(255)
-#  when         :string(255)
-#  map_layer_id :integer
-#  created_at   :datetime         not null
-#  updated_at   :datetime         not null
+#  id                 :integer          not null, primary key
+#  title              :string(255)
+#  latitude           :float
+#  longitude          :float
+#  image_url          :string(255)
+#  image_file_name    :string(255)
+#  image_content_type :string(255)
+#  image_file_size    :string(255)
+#  url                :string(255)
+#  address            :string(255)
+#  when               :string(255)
+#  map_layer_id       :integer
+#  created_at         :datetime         not null
+#  updated_at         :datetime         not null
 #
